pyproject.toml
[project]
name = "twikit"
authors = [
    {name = "d60"},
]
description = "Twitter API wrapper for python with **no API key required**."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "beautifulsoup4",
    "filetype",
    "httpx[socks]",
    "js2py-3-13",
    "lxml",
    "m3u8",
    "pyotp",
    "webvtt-py",
    "xclienttransaction>=1.0.2",
]

dynamic = ["version"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.1",
    "respx>=0.22.0",
    "ruff>=0.15.7",
]

[project.urls]
Homepage = "https://github.com/d60/twikit"
Documentation = "https://twikit.readthedocs.io/en/latest/twikit.html"
Repository = "https://github.com/d60/twikit.git"
Issues = "https://github.com/d60/twikit/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "twikit.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
strict = true
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["live: Test against the actual twitter api with real credentials."]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
